requirements.txt
Django==4.1.4
requests==2.28.2
python-http-client==3.3.7
djangorestframework==3.14.0
beautifulsoup4==4.11.1
html5lib==1.1
django-cors-headers==3.13.0
django-import-export==3.2.0
hashlib